requirement.txt
langchain
langchain_community
faiss-cpu
PyPDF2
streamlit
pypdf
langchain-openai
langchain_openai
ragas